backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
pyyaml>=6.0
gitpython>=3.1.0
httpx>=0.27.0
